SKOJ_BE/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coverage==6.1.2
Django==3.2.9
django-redis==5.0.0
djangorestframework==3.12.4
entrypoints==0.3
flake8==4.0.1
flake8-coding==1.3.2
flake8-quotes==3.3.1
gunicorn==20.1.0
idna==3.3
jsonfield==3.1.0
mccabe==0.6.1
otpauth==1.0.1
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
raven==6.10.0
redis==3.5.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
XlsxWriter==3.0.2
django-dramatiq==0.10.0
dramatiq==1.12.0
django-dbconn-retry==0.1.5
django-cas-ng==4.2.1
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
